+++ CHANGELOG.md
@@ -19,7 +19,7 @@ All notable changes to the "fetch-user-environment" extension will be documented
 ## [1.0.2] - 2017-03-20
 
 - Detect JSON errors with configuration files and alert the user
-- Provide feedback if settings and extensions are up to date when run from the command pallete
+- Provide feedback if settings and extensions are up to date when run from the command palette
 - Extension no longer saves itself when saving the environment
 
 ## [1.0.1] - 2017-03-19